move slog implementation into its own module
